fix funkiness on scrolling down within limited space after inactivity
scrolling downwards results in e.g. (int)(c.y + 0.7) rounding back to c.y, which then in turn hit the case of jumping right to tgt.y, whereas upwards it always rounded away from c.y.
Inactivity & limited space are required to get the delta so small - inactivity results in a couple very fast frames (small gc.deltaNs; this should be fixed but I can't be bothered to figure out how to do it well), and limited space is required for the initial distance to target to be small enough (and the reduced distance to target over time doesn't hit the case as, by the time that gets small enough, gc.deltaNs has returned to normal values)